Trying to get tests to work
