+++ .gitignore
@@ -95,5 +95,5 @@ ENV/
 .pytest_cache/
 
 # project
-*settings.yaml
+/broker_settings.yaml
 inventory.yaml
