--- tests/conftest.py
@@ -1,17 +1,0 @@
-import os
-import pytest
-
-
-@pytest.fixture
-def set_envars(request):
-    """Set and unset one or more environment variables"""
-    if isinstance(request.param, list):
-        for pair in request.param:
-            os.environ[pair[0]] = pair[1]
-        yield
-        for pair in request.param:
-            del os.environ[pair[0]]
-    else:
-        os.environ[request.param[0]] = request.param[1]
-        yield
-        del os.environ[request.param[0]]
