+++ tests/unit/test_settings.py
@@ -4,7 +4,23 @@ import pytest
 from dynaconf import ValidationError
 from broker.exceptions import ConfigurationError
 from broker.providers.test_provider import TestProvider
+import os
+import pytest
 
+
+@pytest.fixture
+def set_envars(request):
+    """Set and unset one or more environment variables"""
+    if isinstance(request.param, list):
+        for pair in request.param:
+            os.environ[pair[0]] = pair[1]
+        yield
+        for pair in request.param:
+            del os.environ[pair[0]]
+    else:
+        os.environ[request.param[0]] = request.param[1]
+        yield
+        del os.environ[request.param[0]]
 
 def test_default_settings():
     test_provider = TestProvider()
